pyproject.toml
[project]
name = "tg-template"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.13"
dependencies = [
    "pytelegrambotapi (>=4.30.0,<5.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "waitress (>=3.0.2,<4.0.0)",
    "update (>=0.0.1,<0.0.2)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.dotenv]
location = ".env.dev"
